# Performance optimizations and fixes