This pull request:

- Resolves errors in the `distanceMostFreqK(between:)` method, and adds additional test cases.
- Refactors the `distanceJaroWinkler(between:)` method to avoid a separate for loop for calculating the number of transpositions.
  - Renames various internal variable names for clarity.
  - Uses `Foundation`s `commonPrefix(with:)` method.
  - I did some local performance analysis on my machine and found that for the modified unit test below the new implementation **reduced the average test run time to 1.28s (whereas current `master` averaged 4.81s), a saving of over 73%.** The `measureMetrics(…)` method calculates the average over 10 runs. N.B. I removed the `XCTAssertEqual` statements for the purposes of benchmarking in order to isolate the performance of the actual method call itself (but I didn't commit this change).
```
func testDistanceJaroWinkler() {
        self.measureMetrics([XCTPerformanceMetric.wallClockTime], automaticallyStartMeasuring: true) {
            for i in 0..<20000 {
                "".distanceJaroWinkler(between: "")
                "".distanceJaroWinkler(between: "Yo")
                "search".distanceJaroWinkler(between: "find")
                "search".distanceJaroWinkler(between: "search")
                "MARTHA".distanceJaroWinkler(between: "MARHTA")
                "DWAYNE".distanceJaroWinkler(between: "DUANE")
                "DIXON".distanceJaroWinkler(between: "DICKSONX")
                "kitten".distanceJaroWinkler(between: "sitting")
                "君子和而不同".distanceJaroWinkler(between: "小人同而不和")
            }
        }
}
```
- Refactors the `distanceHamming(between:)` method for performance. Uses a similar benchmarking method as show above, this **reduced the average test run time to 0.285s (whereas current `master` averaged 0.333s), a saving of over 14%.**
- Uses `String` fast enumeration (rather than subscripting by index, which is very slow) where possible in all string metric methods.
- Avoids repeated calls to `self.count` and `target.count` in all string metric methods (e.g. this should offer small savings to execution time if the methods are called repeatedly in tight loops).
- Corrects some typos in unit test names.