@danielrbrowne 2020-03-02
@autozimu Thanks for accepting the pull request - it might be worth tagging a new release version with these changes?