@autozimu 2020-03-02
Right. I will tag one shortly after.

On Mon, Mar 2, 2020 at 10:37 Daniel Browne <notifications@github.com> wrote:

> @autozimu <https://github.com/autozimu> Thanks for accepting the pull
> request - it might be worth tagging a new release version with these
> changes?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/autozimu/StringMetric.swift/pull/7?email_source=notifications&email_token=AALC3323ULD76Q6S4YTMFC3RFP4HNA5CNFSM4K4HWJG2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOENQOEXA#issuecomment-593551964>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AALC336XSNUHA7LKIKMSEJTRFP4HNANCNFSM4K4HWJGQ>
> .
>