@autozimu 2020-03-02
Released in tag https://github.com/autozimu/StringMetric.swift/releases/tag/0.3.1